Add creation of directory
The directory is not automatically created and thus deploying fails
after reboot.

- kill and rm only if necessary

+++ monit_restart.sh
@@ -23,6 +23,8 @@ JAVA_OPTS="$4"
 
 HOME="/home/sol"
 
+mkdir /tmp/lobid-organisations/ # if rebooted
+
 # it is important to set the proper locale
 . $HOME/.locale
 JAVA_OPTS=$(echo "$JAVA_OPTS" |sed 's#,#\ #g')
@@ -30,8 +32,10 @@ JAVA_OPTS=$(echo "$JAVA_OPTS" |sed 's#,#\ #g')
 cd $HOME/git/$REPO
 case $ACTION in
 	start)
-		kill $(cat target/universal/stage/RUNNING_PID)
-		rm target/universal/stage/RUNNING_PID
+		if [ -f target/universal/stage/RUNNING_PID ]; then
+			kill $(cat target/universal/stage/RUNNING_PID)
+			rm target/universal/stage/RUNNING_PID
+		fi
 		JAVA_OPTS="$JAVA_OPTS -XX:+HeapDumpOnOutOfMemoryError" $HOME/activator-dist-1.3.5/activator "start $PORT"
 		;;
 	stop)
